refactor: Folder organization to handle ports folder.
