+++ app/domain/models/bmodel.py
