+++ app/domain/models/invitation.py
@@ -1,7 +1,7 @@
 from uuid import uuid4
 import secrets
 from datetime import datetime, timezone, timedelta
-from app.domain.bmodel import BModel
+from app.domain.models.bmodel import BModel
 
 
 class Invitation:
