+++ app/domain/ports/email_service_port.py
@@ -1,0 +1,5 @@
+from typing import Protocol
+
+
+class IEmailService(Protocol):
+    def send_email(self, to: str, subject: str, body: str) -> None: ...
