+++ app/domain/ports/ia_service_port.py
@@ -1,0 +1,7 @@
+from typing import Protocol
+
+
+class IAService(Protocol):
+    async def parse_text(self, text: str) -> dict: ...
+
+    async def parse_file(self, file_bytes: bytes, filename: str) -> dict: ...
